azure_monitor/flask_sample/requirements.txt
flask-sqlalchemy >= 2.4.1
flask-wtf >= 0.14.3
opencensus-ext-azure==1.0.8
opencensus-ext-flask==0.7.5
opencensus-ext-logging==0.1.0
opencensus-ext-requests==0.7.5
opencensus-ext-sqlalchemy >= 0.1.2